changed structure, setting up for group graphing
The idea here is first, I wanted to implement some changes to the code base. Second, I am trying to create a system in which group based graphing makes sense. To do this, I've made the luciferase.Config class get and record the number of groups and the size of a group. All groups are assumed to be the same size. Some explanation. A group is a set of plants that are all one type in an experiment. All groups have the same number of members (plants). all members of a group have adjacent indexes and are usually adjacent.

I have also moved most existing files into a folder marked deprecated because I want to change certain things about how the program works.